phonebookReady

+++ OSA2/puhelinluettelo/src/components/PhoneBook.jsx
@@ -6,21 +6,67 @@ const usePhonebook = () => {
   const [newName, setNewName] = useState('')
   const [newNumber, setNewNumber] = useState('')
   const [newFilter, setNewFilter] = useState("")
+  const [errorMessage, setErrorMessage] = useState(null)
+  const [notificationMessage, setNotificationMessage] = useState("info")
 
-  const deletePersons = id => {
-    const changedPersons = persons.filter(p => p.id !== id)
-    PersonsService.remove(id)
-                  .then(personDelete => {
-                    setPersons(changedPersons)
-                  })
+
+
+  const errorMessageFunc = (person, service) => {
+    if (service ==="delete"){
+      setNotificationMessage("info")
+      setErrorMessage(`You have removed ${person}`)
+    }
+    else if (service === "replace") {
+      setNotificationMessage("info")
+      setErrorMessage(`You have replaced ${person}'s number`)
+    }
+    else if (service === "add") {
+      setNotificationMessage("info")
+      setErrorMessage(`You have added ${person}`)
+    }
+    else if (service === "deleteWarning") {
+      setNotificationMessage("error")
+      setErrorMessage(`Information of ${person} has already been removed from server`)
+    }
+
+    setTimeout(() => {
+      setErrorMessage(null)
+    }, 3000)
   }
 
-  useEffect(() =>{
-    PersonsService.getAll()
-                  .then(initialPersons => {
-                    return setPersons(initialPersons)
-                  })
-    }, [])
+  const deletePersons = id => {
+    if (window.confirm(`Delete ${persons.find(p=> p.id === id).name} ?`)){
+      const changedPersons = persons.filter(p => p.id !== id)
+      PersonsService.remove(id)
+                    .then(personDelete => {
+                      setPersons(changedPersons)
+                    })
+                    .catch(error => {
+                      errorMessageFunc(persons.find(p=> p.id === id).name,"deleteWarning")
+                    })
+      errorMessageFunc(persons.find(p=> p.id === id).name, "delete")
+                  }
+  }
+
+  const Notification = ({message }) => {
+    if (message === null) {
+      return null
+    }
+    else {
+      if (notificationMessage === "error")
+      {  return (
+          <div className="error">
+            {message}
+          </div>)}
+      else if (notificationMessage === "info") {
+        return (
+          <div className="info">
+            {message}
+          </div>
+      )}   
+      }
+  }
+
 
   const addPerson = (event) => {
       event.preventDefault()
@@ -29,19 +75,33 @@ const usePhonebook = () => {
         number: newNumber,
       }
       const sameName = personObject.name.toUpperCase()
-      //console.log('lisättävä objekti: ', personObject)
-      if (persons.map(p => p.name.toUpperCase()).includes(sameName)) 
-        alert(`${personObject.name} is already added to phonebook`)
+      const ppl = persons.find(p=> p.name === personObject.name)
+      if (persons.map(p => p.name.toUpperCase()).includes(sameName)){ 
+        if (window.confirm(`${ppl.name} is already added to phonebook, replace old number with new one?`)) 
+          {PersonsService
+            .update(ppl.id,personObject)
+            .then(response =>{
+                setPersons(newPersons =>
+                  newPersons.concat(response.data))
+                setNewName('')
+                setNewNumber('')
+            })
+            .catch(error =>
+              {
+                errorMessageFunc(ppl.name,"deleteWarning")
+              }
+            )
+          errorMessageFunc(ppl.name,"replace")}
+        else
+          alert(`${personObject.name} is already added to phonebook`)
+        }
       else 
-        PersonsService
+        {PersonsService
             .create(personObject)
             .then(response => 
               {
-              const rData = response.data
               setPersons(prevPersons =>
-                prevPersons.concat(rData))
-              //console.log('Phonebookin response data:', response)
-              //console.log('PhoneBookin persons: ', persons)
+                prevPersons.concat(response.data))
               setNewName('')
               setNewNumber('')
               
@@ -50,6 +110,8 @@ const usePhonebook = () => {
               {
               console.log('Error: ', error)
             })
+
+        errorMessageFunc(personObject.name, "add")}
       }
 
   const handleNameChange = (event) => {
@@ -70,7 +132,9 @@ const usePhonebook = () => {
     handleNumberChange,
     newFilter,
     setNewFilter,
-    deletePersons
+    deletePersons,
+    Notification,
+    errorMessage
   }
 }
